+++ omechu-app/src/app/auth/find-password/page.tsx
@@ -8,7 +8,7 @@ import { useRouter } from "next/navigation";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useForm } from "react-hook-form";
 
-import Input from "@/app/components/auth/Input";
+import Input from "@/app/auth/components/Input";
 import SquareButton from "@/app/components/common/button/SquareButton";
 import {
   findPasswordSchema,
