+++ omechu-app/src/app/auth/login/page.tsx
@@ -8,7 +8,7 @@ import Link from "next/link";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useForm } from "react-hook-form";
 
-import Input from "@/app/components/auth/Input";
+import Input from "@/app/auth/components/Input";
 import SquareButton from "@/app/components/common/button/SquareButton";
 import { loginSchema, LoginFormValues } from "@/lib/schemas/auth.schema";
